Conflicts:
	buddy/upk_buddy.c
		- extra DEBUG
	controller/controller.c
		- clobbered change to async

+++ controller/upk.c
@@ -166,8 +166,8 @@ int cmdline_join(
         }
 
         if( cp - buf + strlen( argv[i] ) > buf_len - 2 ) {
-            printf("Buffer overflow in cmdline_join (%ld > %d)\n",
-                    cp - buf + strlen( argv[i] ), buf_len - 2 );
+            printf("Buffer overflow in cmdline_join (%d > %d)\n",
+                    (int) (cp - buf + strlen( argv[i] )), buf_len - 2 );
             return( -1 );
         }
         if( cp != buf ) {
